Refactor MultistageStopwatch implementations.
Introduce AbstractMultistageStopwatch and let SimpleMultistageStopwatch
and HistogramBasedStopwatch inherit most behavior from it.

Eliminate SimpleMultistageTimer because it's obsoleted by
ListenableTimer. AbstractMultistageStopwatch, SimpleMultistageStopwatch,
and HistogramBasedStopwatch now just use a ListenableTimer listener to
do all their timer-stopped housekeeping.

Adjust BlockingExecutorServiceTest to not depend on the old startTimer
method that returned the timer, and to not use so many mocks. (It now
uses a plain old test-helper subclass.)

Add real equals & hashCode methods to  SimpleMultistageStopwatch and
HistogramBasedStopwatch.